+++ kit/webpack/browser.js
@@ -86,8 +86,8 @@ export default new WebpackConfig().extend('[root]/base.js').merge({
     new webpack.optimize.CommonsChunkPlugin({
       name: 'vendor',
       minChunks: module => (
-         // this assumes your vendor imports exist in the node_modules directory
-         module.context && module.context.indexOf('node_modules') !== -1
+        // this assumes your vendor imports exist in the node_modules directory
+        module.context && module.context.indexOf('node_modules') !== -1
       ),
     }),