+++ src/app.js
@@ -1,197 +1,91 @@
+// Your app's entry point.  Every ReactQL projects requires 'src/app.js',
+// which both the server and browser will import.
+//
+// In this file, you'll do two things:
+//
+// 1.  Import `kit/config`, and configure your app.  In this example, I'm
+// adding a custom Redux reducer that acts as a simple counter, and enabling
+// a built-in GraphQL server that imports a schema for a simple message.
+//
+// 2.  Export the root React component that goes between <div id="main"/>
+// in the server-side HTML.
+
 // ----------------------
 // IMPORTS
 
-/* NPM */
+/* ReactQL */
 
-// React
-import React from 'react';
-import PropTypes from 'prop-types';
+// Config API, for adding reducers and configuring our ReactQL app
+import config from 'kit/config';
 
-// GraphQL
-import { graphql } from 'react-apollo';
+/* App */
 
-// Listen to Redux store state
-import { connect } from 'react-redux';
+// Example counter reducer.  This simply increments the counter by +1
+import counterReducer from 'src/reducers/counter';
 
-// Routing
-import {
-  Link,
-  Route,
-  Switch,
-} from 'react-router-dom';
+// Main component -- i.e. the 'root' React component in our app
+import Main from 'src/components/main';
 
-// <Helmet> component for setting the page title
-import Helmet from 'react-helmet';
-
-/* Local */
-
-// NotFound 404 handler for unknown routes
-import { NotFound, Redirect } from 'kit/lib/routing';
-
-// GraphQL queries
-import allMessages from 'src/queries/all_messages.gql';
-
-// Styles
+// Init global styles.  These will be added to the resulting CSS automatically
+// without any class hashing.  Use this to include default or framework CSS.
 import './styles.global.css';
-import css from './styles.css';
-import sass from './styles.scss';
-import less from './styles.less';
-
-// Get the ReactQL logo.  This is a local .svg file, which will be made
-// available as a string relative to [root]/dist/assets/img/
-import logo from './reactql-logo.svg';
 
 // ----------------------
 
-// We'll display this <Home> component when we're on the / route
-const Home = () => (
-  <h1>You&apos;re on the home page - click another link above</h1>
-);
+/* REDUCERS */
 
-// Helper component that will be conditionally shown when the route matches.
-// This gives you an idea how React Router v4 works
-const Page = ({ match }) => (
-  <h1>Changed route: {match.params.name}</h1>
-);
+// Add our custom `counter` reducer.  This must follow the shape
+// {state, reducer()}, otherwise it will throw an error
+config.addReducer('counter', counterReducer);
 
-// Create a route that will be displayed when the code isn't found
-const WhenNotFound = () => (
-  <NotFound>
-    <h1>Unknown route - the 404 handler was triggered!</h1>
-  </NotFound>
-);
+// Set our server config, by checking `SERVER` -- this code path will be
+// eliminated by Webpack in the browser, so we can safely add this.
 
-// Specify PropTypes if the `match` object, which is injected to props by
-// the <Route> component
-Page.propTypes = {
-  match: PropTypes.shape({
-    params: PropTypes.object,
-  }).isRequired,
-};
+if (SERVER) {
+  /* GRAPHQL */
 
-// Stats pulled from the environment.  This demonstrates how data will
-// change depending where we're running the code (environment vars, etc)
-// and also how we can connect a 'vanilla' React component to an RxJS
-// observable source, and feed eventual values in as properties
-const Stats = () => {
-  const info = [
-    ['Environment', process.env.NODE_ENV],
-  ];
+  // If we're running on the server, create a built-in GraphQL server using
+  // custom schema.  By default, this will bind to /graphql -- POST will
+  // handle GraphQL requests; GET will display the GraphiQL query interface
+  //
+  // Apollo will attempt to connect to /graphql automatically.  We'll run
+  // this only on the server, to avoid the bloat of loading unnecessary GraphQL
+  // schema and types on the client
+  config.enableGraphQLServer(require('src/graphql/schema').default);
 
-  return (
-    <ul className={css.data}>
-      {info.map(([key, val]) => (
-        <li key={key}>{key}: <span>{val}</span></li>
-      ))}
-    </ul>
-  );
-};
+  /* CUSTOM ROUTES */
 
-// Now, let's create a GraphQL-enabled component...
+  // We can add custom routes to the web server easily, by using
+  // `config.add<Get|Post|Put|Patch>Route()`.  Note:  These are server routes only.
+  config.addGetRoute('/test', async ctx => {
+    ctx.body = 'Hello from your ReactQL route.';
+  });
 
-// ... then, let's create the component and decorate it with the `graphql`
-// HOC that will automatically populate `this.props` with the query data
-// once the GraphQL API request has been completed
-@graphql(allMessages)
-class GraphQLMessage extends React.PureComponent {
-  static propTypes = {
-    data: PropTypes.shape({
-      allMessages: PropTypes.arrayOf(
-        PropTypes.shape({
-          text: PropTypes.string.isRequired,
-        }),
-      ),
-    }),
-  }
+  /* CUSTOM 404 HANDLER */
 
-  render() {
-    const { data } = this.props;
-    const message = data.allMessages && data.allMessages[0].text;
-    const isLoading = data.loading ? 'yes' : 'nope';
-    return (
-      <div>
-        <h2>Message from GraphQL server: <em>{message}</em></h2>
-        <h2>Currently loading?: {isLoading}</h2>
-      </div>
-    );
-  }
+  // By default, if the server gets a route request that results in a 404,
+  // it will set `ctx.status = 404` but continue to render the <NotFound>
+  // block as normal.  If we want to add our own custom handler, we can use
+  // `config.set404Handler()` as below.
+  //
+  // Note:  This only applies to SERVER routes.  On the client, the
+  // <NotFound> block will *always* run.
+
+  config.set404Handler((ctx, store) => {
+    // For demo purposes, let's get a JSON dump of the current Redux state
+    // to see that we can expect its contents
+    const stateDump = JSON.stringify(store.getState());
+
+    // Explicitly set the return status to 404.  This is done for us by
+    // default if we don't have a custom 404 handler, but left to the function
+    // otherwise (since we might not always want to return a 404)
+    ctx.status = 404;
+
+    // Set the body
+    ctx.body = `This route does not exist on the server - Redux dump: ${stateDump}`;
+  });
 }
 
-// Example of CSS, SASS and LESS styles being used together
-const Styles = () => (
-  <ul className={css.styleExamples}>
-    <li className={css.example}>Styled by CSS</li>
-    <li className={sass.example}>Styled by SASS</li>
-    <li className={less.example}>Styled by LESS</li>
-  </ul>
-);
-
-// Sample component that demonstrates using a part of the Redux store
-// outside of Apollo.  We can import own custom reducers in `kit/lib/redux`
-// and 'listen' to them here
-@connect(state => ({ counter: state.counter }))
-class ReduxCounter extends React.PureComponent {
-  static propTypes = {
-    counter: PropTypes.shape({
-      count: PropTypes.number.isRequired,
-    }),
-  };
-
-  // Trigger the `INCREMENT_COUNTER` action in Redux, to add 1 to the total
-  triggerIncrement = () => {
-    this.props.dispatch({
-      type: 'INCREMENT_COUNTER',
-    });
-  }
-
-  render() {
-    const { count } = this.props.counter;
-    return (
-      <div>
-        <h2>Listening to Redux counter: {count}</h2>
-        <button onClick={this.triggerIncrement}>Increment</button>
-      </div>
-    );
-  }
-}
-
-// Export a simple component that allows clicking on list items to change
-// the route, along with a <Route> 'listener' that will conditionally display
-// the <Page> component based on the route name
-export default () => (
-  <div>
-    <Helmet
-      title="ReactQL application"
-      meta={[{
-        name: 'description',
-        content: 'ReactQL starter kit app',
-      }]} />
-    <div className={css.hello}>
-      <img src={logo} alt="ReactQL" className={css.logo} />
-    </div>
-    <hr />
-    <GraphQLMessage />
-    <hr />
-    <ul>
-      <li><Link to="/">Home</Link></li>
-      <li><Link to="/page/about">About</Link></li>
-      <li><Link to="/page/contact">Contact</Link></li>
-      <li><Link to="/old/path">Redirect from /old/path &#8594; /new/path</Link></li>
-    </ul>
-    <hr />
-    <Switch>
-      <Route exact path="/" component={Home} />
-      <Route path="/page/:name" component={Page} />
-      <Redirect from="/old/path" to="/new/path" />
-      <Route component={WhenNotFound} />
-    </Switch>
-    <hr />
-    <ReduxCounter />
-    <hr />
-    <p>Runtime info:</p>
-    <Stats />
-    <hr />
-    <p>Stylesheet examples:</p>
-    <Styles />
-  </div>
-);
+// In app.js, we need to export the root component we want to mount as the
+// starting point to our app.  We'll just export the `<Main>` component.
+export default Main;